+++ STM32/app/src/main.c
@@ -16,9 +16,6 @@
 
 static void SystemClock_Config();
 
-#define RECEIVE_TIMEOUT 	10000
-#define LED_PACKET_SIZE		1
-
 int main()
 {
 	// Initialize System clock to 48MHz from external clock
@@ -39,132 +36,6 @@ int main()
 
 	// Boucle de contrôle de subordonné
 	subordonneMain();
-}
-
-/* CODE QUI GERE LE SUBORDONNE */
-/* Attend en boucle un message de la part du contrôleur et renvoie un paquet d'acknowledge ou de
- * non acknowledge en fonction du message reçu */
-int subordonneMain(void)
-{
-	frameSACS_s packetLed = {CID, ACK, LED_PACKET_SIZE, {LED_TOGGLE}, 0};
-	frameSACS_s receivedPacket;
-	uint8_t receiveStatus;
-	uint8_t sendStatus;
-
-	// Initialize System clock to 48MHz from external clock
-	SystemClock_Config();
-	// Initialize timebase
-	BSP_TIMER_Timebase_Init();
-	// Initialize NVIC
-	BSP_NVIC_Init();
-	// Initialize SPI interface
-	BSP_SPI1_Init();
-	// Initialize Debug Console
-	BSP_Console_Init();
-
-	my_printf("Console ready!\r\n");
-
-	///////////////////////////////////////////
-	//setup SX1272
-	APP_SX1272_setup();
-    // BSP_LED_Init();
-
-	while(1)
-	{
-        /* Receive data from controller */
-		receiveStatus = APP_SACS_receiveSub(&receivedPacket, RECEIVE_TIMEOUT, SID);
-		switch (receiveStatus)
-		{
-		case RECEIVE_OK:
-			#if DEBUG
-				my_printf("Receive OK\r\n");
-			#endif
-			/* Verification de l'action à réaliser */
-            switch (receivedPacket.data[0]);
-            {
-                case LED_ON:
-                    // BSP_LED_On();
-					packetLed.ack = ACK;
-                    break;
-                case LED_OFF:
-                    // BSP_LED_Off();
-					packetLed.ack = ACK;
-                    break;
-                case LED_TOGGLE:
-                    // BSP_LED_Toggle();
-					packetLed.ack = ACK;
-                    break;
-            }
-			break;
-
-		case RECEIVE_ERROR:
-			#if DEBUG
-				my_printf("Receive ERROR\r\n");
-			#endif
-            packetLed.ack = NACK;
-			break;
-
-		case RECEIVE_FAILED:
-			#if DEBUG
-				my_printf("Receive FAIL\r\n");
-			#endif
-			packetLed.ack = NACK;
-			break;
-
-		case CRC_ERROR :
-			#if DEBUG
-				my_printf("CRC ERROR\r\n");
-			#endif
-			packetLed.ack = NACK;
-			break;
-
-		case SIZE_ERROR :
-			#if DEBUG
-				my_printf("SIZE ERROR\r\n");
-			#endif
-			packetLed.ack = NACK;
-			break;
-
-		case RECEIVE_SUB_NC :
-			#if DEBUG
-				my_printf("Receive but not for me\r\n");
-			#endif
-			break;
-
-		default:
-			#if DEBUG
-				my_printf("Unmanaged receive error\r\n");
-			#endif
-            packetLed.ack = NACK;
-			break;
-		}
-
-		if (receiveStatus != RECEIVE_SUB_NC)
-		{
-			/* Send ACK packet */
-			BSP_DELAY_ms(1000); // Laisse un peu le temps
-			sendStatus = APP_SACS_send(packetLed);
-			if(sendStatus != SEND_OK)
-			{
-				#if DEBUG
-					my_printf("Send ERROR\r\n");
-				#endif
-				return SEND_ERROR;
-			}
-			else
-			{
-				#if DEBUG
-					my_printf("Send OK\r\n");
-				#endif
-			}
-		}
-		else
-		{
-			// rien du tout
-		}
-	}
-	/* Should never go there */
-	return EXIT_FAILURE;
 }
 
 /*
